tidied code
