+++ Makefile
@@ -12,3 +12,5 @@ test-pc:
 	go test -v -cover ./problems/...
 vet:
 	go vet ./...
+build:
+	go build -o ./bin/cli ./cli/main.go
